requirements.txt
streamlit==1.37.1
pypdf==4.3.1
faiss-cpu==1.13.2

sentence-transformers==3.0.1
numpy==2.1.0

pandas==2.2.2

plotly==5.23.0
python-dotenv==1.0.1
tiktoken==0.7.0

scikit-learn==1.5.1
umap-learn==0.5.6

fastapi==0.112.0
uvicorn==0.30.6

google-genai==1.0.0
